// src/models/public/Occurrences.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { ObservationImportsObservationImportId } from './ObservationImports';
import type { PlacesPlaceId } from './Places';

/** Identifier type for public.occurrences */
export type OccurrencesOccurrenceId = string & { __brand: 'public.occurrences' };

/**
 * Represents the table public.occurrences
 * GBIF occurrences. Imported for subprojects (species projects) or projects (biotope projects).
 */
export default interface Occurrences {
  occurrence_id: OccurrencesOccurrenceId;

  account_id: AccountsAccountId | null;

  observation_import_id: ObservationImportsObservationImportId | null;

  /** The place this occurrence is assigned to. */
  place_id: PlacesPlaceId | null;

  not_to_assign: boolean | null;

  comment: string | null;

  /** data as received from GBIF */
  data: unknown | null;

  /** Used to replace previously imported occurrences */
  id_in_source: string | null;

  /** geometry of occurrence. Extracted from data to show the occurrence on a map */
  geometry: unknown | null;

  /** label of occurrence, used to show it in the UI. Created on import */
  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.occurrences
 * GBIF occurrences. Imported for subprojects (species projects) or projects (biotope projects).
 */
export interface OccurrencesInitializer {
  /** Default value: uuid_generate_v7() */
  occurrence_id?: OccurrencesOccurrenceId;

  account_id?: AccountsAccountId | null;

  observation_import_id?: ObservationImportsObservationImportId | null;

  /** The place this occurrence is assigned to. */
  place_id?: PlacesPlaceId | null;

  /** Default value: false */
  not_to_assign?: boolean | null;

  comment?: string | null;

  /** data as received from GBIF */
  data?: unknown | null;

  /** Used to replace previously imported occurrences */
  id_in_source?: string | null;

  /** geometry of occurrence. Extracted from data to show the occurrence on a map */
  geometry?: unknown | null;

  /** label of occurrence, used to show it in the UI. Created on import */
  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.occurrences
 * GBIF occurrences. Imported for subprojects (species projects) or projects (biotope projects).
 */
export interface OccurrencesMutator {
  occurrence_id?: OccurrencesOccurrenceId;

  account_id?: AccountsAccountId | null;

  observation_import_id?: ObservationImportsObservationImportId | null;

  /** The place this occurrence is assigned to. */
  place_id?: PlacesPlaceId | null;

  not_to_assign?: boolean | null;

  comment?: string | null;

  /** data as received from GBIF */
  data?: unknown | null;

  /** Used to replace previously imported occurrences */
  id_in_source?: string | null;

  /** geometry of occurrence. Extracted from data to show the occurrence on a map */
  geometry?: unknown | null;

  /** label of occurrence, used to show it in the UI. Created on import */
  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}